Change the file name
1. Change the filename, better understand each file's role
2. Improve the front-end, fix some bugs + improve some functionality
