+++ src/main/java/interfaceAdaptors/servletDelete.java
@@ -1,4 +1,4 @@
-package com.xxxx.example;
+package interfaceAdaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -6,11 +6,12 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-// The servletLikePostResponseLike class works as a presenter and pass like message back to front-end.
-@WebServlet("/likePostResponseLike")
-public class servletLikePostResponseLike extends HttpServlet {
+
+// The servletDelete class works as a controller and pass variables to gateway.
+@WebServlet("/listenDelete")
+public class servletDelete extends HttpServlet {
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("succeed to like");
+        req.getRequestDispatcher("/deleteGateway").forward(req,resp);
     }
 }
