+++ src/main/java/interfaceAdaptors/servletLikePostResponseDislike.java
@@ -1,4 +1,4 @@
-package com.xxxx.example;
+package interfaceAdaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,12 +7,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletSendReplyResponse class works as a presenter and pass message back to front-end.
-@WebServlet("/sendReplyResponse")
-public class servletSendReplyResponse extends HttpServlet {
+// The servletLikePostResponseDislike class works as a presenter and pass dislike message back to front-end.
+@WebServlet("/likePostResponseDislike")
+public class servletLikePostResponseDislike extends HttpServlet {
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.setCharacterEncoding("UTF-8");
-        resp.getWriter().append("success");
+        resp.getWriter().append("succeed to dislike");
     }
 }
