+++ src/main/java/interfaceAdaptors/servletLikePostResponseLike.java
@@ -1,4 +1,4 @@
-package com.xxxx.example;
+package interfaceAdaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -6,12 +6,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-
-// The servletDelete class works as a controller and pass variables to gateway.
-@WebServlet("/listenDelete")
-public class servletDelete extends HttpServlet {
+// The servletLikePostResponseLike class works as a presenter and pass like message back to front-end.
+@WebServlet("/likePostResponseLike")
+public class servletLikePostResponseLike extends HttpServlet {
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("/deleteGateway").forward(req,resp);
+        resp.getWriter().append("succeed to like");
     }
 }
