+++ src/main/java/interfaceAdaptors/servletLoginResponsePasswordIncorrect.java
@@ -1,4 +1,4 @@
-package com.xxxx.example;
+package interfaceAdaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,11 +7,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletLikePostResponseDislike class works as a presenter and pass dislike message back to front-end.
-@WebServlet("/likePostResponseDislike")
-public class servletLikePostResponseDislike extends HttpServlet {
+// The servletLoginResponsePasswordIncorrect class works as a presenter and pass incorrect message back to front-end.
+@WebServlet("/loginResponsePasswordIncorrect")
+public class servletLoginResponsePasswordIncorrect extends HttpServlet {
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("succeed to dislike");
+        resp.getWriter().append("incorrect");
     }
 }
