+++ src/main/java/interfaceAdaptors/servletLoginResponseUserNotExist.java
@@ -1,4 +1,4 @@
-package com.xxxx.example;
+package interfaceAdaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,10 +7,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletRegisterResponseUserExist class works as a presenter and pass invalid message back to front-end.
-@WebServlet("/registerResponseUserExist")
-public class servletRegisterResponseUserExist extends HttpServlet {
+// The servletLoginResponseUserNotExist class works as a presenter and pass user not exist message back to front-end.
+@WebServlet("/loginResponseUserNotExist")
+public class servletLoginResponseUserNotExist extends HttpServlet {
+    @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("username already exist");
+        resp.getWriter().append("not_exist");
     }
 }
