+++ src/main/java/interfaceAdaptors/servletPostPost.java
@@ -1,4 +1,6 @@
-package com.xxxx.example;
+package interfaceAdaptors;
+
+import database_connection.Database;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,10 +9,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletRegisterResponsePasswordInvalid class works as a presenter and pass incorrect message back to front-end.
-@WebServlet("/registerResponsePasswordInvalid")
-public class servletRegisterResponsePasswordInvalid extends HttpServlet {
+// The servletPostPost class works as a controller and pass variables to gateway.
+@WebServlet("/listenPostPost")
+public class servletPostPost extends HttpServlet {
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("password not match");
+        req.setCharacterEncoding("UTF-8");
+        req.getRequestDispatcher("/postPostGateway").forward(req,resp);
     }
 }
+
