+++ src/main/java/interfaceAdaptors/servletPostPostResponse.java
@@ -1,4 +1,6 @@
-package com.xxxx.example;
+package interfaceAdaptors;
+
+import database_connection.Database;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,11 +9,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletRegisterResponse class works as a presenter and pass success message back to front-end.
-@WebServlet("/registerResponse")
-public class servletRegisterResponse extends HttpServlet {
+// The servletPostPostResponse class works as a presenter and pass message back to front-end.
+@WebServlet("/postPostResponse")
+public class servletPostPostResponse extends HttpServlet {
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         resp.getWriter().append("success");
     }
 }
-
