+++ src/main/java/interfaceAdaptors/servletRegisterResponse.java
@@ -1,4 +1,4 @@
-package com.xxxx.example;
+package interfaceAdaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,11 +7,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletLoginResponsePasswordIncorrect class works as a presenter and pass incorrect message back to front-end.
-@WebServlet("/loginResponsePasswordIncorrect")
-public class servletLoginResponsePasswordIncorrect extends HttpServlet {
-    @Override
+// The servletRegisterResponse class works as a presenter and pass success message back to front-end.
+@WebServlet("/registerResponse")
+public class servletRegisterResponse extends HttpServlet {
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("incorrect");
+        resp.getWriter().append("success");
     }
 }
+
