+++ src/main/java/interfaceAdaptors/servletRegisterResponseUserExist.java
@@ -1,4 +1,4 @@
-package com.xxxx.example;
+package interfaceAdaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,11 +7,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletLoginResponseSuccess class works as a presenter and pass success message back to front-end.
-@WebServlet("/loginResponseSuccess")
-public class servletLoginResponseSuccess extends HttpServlet {
-    @Override
+// The servletRegisterResponseUserExist class works as a presenter and pass invalid message back to front-end.
+@WebServlet("/registerResponseUserExist")
+public class servletRegisterResponseUserExist extends HttpServlet {
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("success");
+        resp.getWriter().append("username already exist");
     }
 }
