+++ src/main/java/interfaceAdaptors/servletSendReplyResponse.java
@@ -1,4 +1,4 @@
-package com.xxxx.example;
+package interfaceAdaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,12 +7,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// Not yet functioning (to be added later)
-@WebServlet("/likeReplyResponse")
-public class servletLikeReplyResponse extends HttpServlet {
+// The servletSendReplyResponse class works as a presenter and pass message back to front-end.
+@WebServlet("/sendReplyResponse")
+public class servletSendReplyResponse extends HttpServlet {
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String status = (String) req.getAttribute("status");
-        resp.getWriter().append(status);
+        resp.setCharacterEncoding("UTF-8");
+        resp.getWriter().append("success");
     }
 }
